requirements.txt
setuptools
psycopg2-binary
starlette==0.12.9
uvicorn==0.10.9
SQLAlchemy==1.3.12
pydantic==1.2
fastapi==0.45.0
python-dotenv==0.10.3
gunicorn